==> notebooks/01_sftl_scrape.ipynb <==
# Scraps from the Loft
# Stand Up Comedy Script Scrape

## Imports

import requests
from bs4 import BeautifulSoup as BS
import numpy as np
import pandas as pd
import re

## Scrape

endpoint = 'https://scrapsfromtheloft.com/stand-up-comedy-scripts/'
response = requests.get(endpoint, 'lxml')
soup = BS(response.content)

toc = soup.find_all('h3')[4:]

transcript_hrefs = [item.a['href'] for item in toc]
transcript_descriptions = [item.a.contents[0].strip() for item in toc]

# Construct loop to only pull out transcript content.
transcripts = []
for url in transcript_hrefs:
    response = requests.get(url)
    soup = BS(response.content)
    p_tags = soup.find_all('p', attrs={'style':'text-align: justify;'})
    transcript = " ".join([p.text for p in p_tags])
    transcripts.append(transcript)

transcripts_df = pd.DataFrame({
    'description': transcript_descriptions,
    'link': transcript_hrefs,
    'transcript':transcripts
})

transcripts_df.to_pickle('../data/transcripts_raw_df.pickle')
