==> notebooks/02_imdb_api.ipynb <==
# Get Standup Show Ratings

# IMBD API

## Imports

import requests
import numpy as np
import pandas as pd
import re
import json
from datetime import date

import pickle

import plotly.express as px

tscripts = pd.read_pickle('../data/transcripts_raw_df.pickle')
tscripts['script characters'] = tscripts['transcript'].apply(len)
px.strip(tscripts, x = 'script characters', hover_data= ['description'])


tscripts = tscripts[tscripts['script characters'] > 21000]
shows_split = tscripts['description'].str.split('|', expand = True)[0]

def transcript_strip(string):
    hasyear = re.search(r"\(\d{4}\)", string)
    hasdashbracket = re.search(r"\s[–\[]", string)
    if hasyear:
        return string[:hasyear.end()]
    elif hasdashbracket:
        return string[:hasdashbracket.start()]
    else:
        return string

show_searches =  [transcript_strip(search) for search in shows_split]

## Get IDs from API
I'll have to only do 100 results per day since I'm sticking with the free version of the API.
After I get all the title IDs (4 days), I'll then have to get the ratings (another 4 days). If I want the reviews, that's another 4 days.

api_key = json.load(open('../../../api_keys/api_keys.json'))['imdb2']

# with open(f'../data/imdb_title_search_results_chris_2022-05-23.pickle', 'rb') as file:
#     search_results = pickle.load(file)

# with open(f'../data/imdb_title_search_results_fails_chris_2022-05-23.pickle', 'rb') as file:
#     search_results_fails = pickle.load(file)

start = None
end = None
for search in show_searches[start:end]:

    try:
        endpoint = 'https://imdb-api.com/en/API/SearchTitle/'
        params = {
            'apiKey':api_key,
            'expression':search
        }
        response = requests.get(endpoint, params)
        result = response.json()

        search_result = result['results'][0]
        search_results.append(search_result)
    except:
        search_results_fails.append(search)

# with open(f'../data/imdb_title_search_results_{date.today()}.pickle', 'wb') as file:
#     pickle.dump(search_results, file)

# with open(f'../data/imdb_title_search_results_fails_{date.today()}.pickle', 'wb') as file:
#     pickle.dump(search_results_fails, file)

## Get Title Info from API

# title_results = []

# title_results_fails = []
# with open(f'../data/imdb_title_results_{date.today()}.pickle', 'rb') as file:
#     title_results = pickle.load(file)

start = 299
end = None
for search_dict in search_results[start:end]:

    try:
        endpoint = 'https://imdb-api.com/en/API/Title/'
        params = {
            'apiKey': api_key,
            'id': search_dict['id']
        }
        response = requests.get(endpoint, params)
        title_result = response.json()

        title_results.append(title_result)
    except:
        title_results_fails.append(search_dict['title'])

# with open(f'../data/imdb_title_results_{date.today()}.pickle', 'wb') as file:
#     pickle.dump(title_results, file)

## Get Reviews from API
Unfortunately, these appear capped at 25.

endpoint = 'https://imdb-api.com/en/API/Reviews/'
params = {
    'apiKey': api_key,
    'id': search_results[0]['id']
}
response = requests.get(endpoint, params)
result = response.json()
result

# review_results_fails = []
# review_results = []

start = 279
end = None
for search_dict in search_results[start:end]:

    try:
        endpoint = 'https://imdb-api.com/en/API/Reviews/'
        params = {
            'apiKey': api_key,
            'id': search_dict['id']
        }
        response = requests.get(endpoint, params)
        review_result = response.json()

        review_results.append(review_result)
    except:
        review_results_fails.append(search_dict['title'])

len(review_results)

with open(f"../data/imdb_review_results_{date.today()}", 'wb') as file:
    pickle.dump(review_results, file)

# max number of reviews is 25. this must be capped.
max([len(rev_res['items']) for rev_res in review_results])

## Try those short titles

tscripts = pd.read_pickle('../data/transcripts_raw_df.pickle')
tscripts['script characters'] = tscripts['transcript'].apply(len)
short_tscripts = tscripts[tscripts['script characters'] < 21000]

short_shows_split = short_tscripts['description'].str.split('|', expand = True)[0]
short_show_searches =  [transcript_strip(search) for search in short_shows_split]

# short_search_results = []
# short_search_results_fails = []

start = None
end = 57
for search in short_show_searches[start:end]:

    try:
        endpoint = 'https://imdb-api.com/en/API/SearchTitle/'
        params = {
            'apiKey':api_key,
            'expression':search
        }
        response = requests.get(endpoint, params)
        result = response.json()

        short_search_result = result['results'][0]
        short_search_results.append(short_search_result)
    except:
        short_search_results_fails.append(search)

short_title_results = []
short_title_results_fails = []

for search_dict in short_search_results:

    try:
        endpoint = 'https://imdb-api.com/en/API/Title/'
        params = {
            'apiKey': api_key,
            'id': search_dict['id']
        }
        response = requests.get(endpoint, params)
        short_title_result = response.json()

        short_title_results.append(short_title_result)
    except:
        short_title_results_fails.append(search_dict['title'])

## Reconciliation

empty_search_dict = dict(zip(title_results[0].keys(), [None for i in range(len(title_results[0].keys()))]))

for entry in short_search_results_fails:
    ind = short_show_searches.index(entry)
    short_title_results.insert(ind, empty_search_dict)

full_title_results = title_results + short_title_results
for entry in search_results_fails:
    ind = show_searches.index(entry)
    full_title_results.insert(ind, empty_search_dict)

len(short_title_results)

len(full_title_results)

with open(f"../data/imdb_full_title_results_{date.today()}.pickle", 'wb') as file:
    pickle.dump(full_title_results, file)

tscripts_df_cut = pd.concat([tscripts[tscripts['script characters'] > 21000], tscripts[tscripts['script characters'] < 21000]]).reset_index(drop = True)
tscripts_df_cut

with open(f"../data/transcripts_df_{date.today()}.pickle", 'wb') as file:
    pickle.dump(tscripts_df_cut, file)
