==> notebooks/07_create_community_dfs.ipynb <==
# Community Analysis

## Table of Contents

   1. [imports](#imports)
   2. [make dataframes](#make-dataframes)

## imports

import pandas as pd
import numpy as np
import pickle
import plotly.express as px

import re

from sklearn.decomposition import PCA
from sklearn.preprocessing import StandardScaler, OneHotEncoder
from sklearn.compose import ColumnTransformer
from sklearn.pipeline import Pipeline
from umap import UMAP
from sklearn.decomposition import PCA

with open(f'../data/metascripts_df_sentiment.pickle', 'rb') as file:
    metascripts = pickle.load(file)

with open('../data/features_dict.pickle', 'rb') as file:
    features_dict = pickle.load(file)

with open('../data/feature_cosims_dict.pickle', 'rb') as file:
    feature_cosims_dict = pickle.load(file)

## make dataframes

   - descom_df: show descriptions and mean communities from Louvain algorithm
   - all_features: each show pair and all feature-group similarities
   - metacosims: all_features + metadata
   - graph_all_features: all_features deduplicated for graph database upload
   - feature_mean_coms: each feature group with description (show) labels and corresponding mean communities from Louvain algorithm

### descom_df: descriptions and communities

descriptions = [re.search('(?<=\":).+(?=,\")', string).group(0) for string in pd.read_csv('../data/graph_coms_df.csv')['s'].values]
communities = [int(re.search('(?<=\"meanCommunity\":)\d+', string).group(0)) for string in pd.read_csv('../data/graph_coms_df.csv')['s'].values]
descom_df = pd.DataFrame({
        'description': descriptions,
        'community':communities
    })

### all_features: each show pair and all feature-group similarities

# create the all_features dataframe by
# melting each cosims dictionary and joining them together on the descriptions
all_features = pd.DataFrame()
for feature_group in feature_cosims_dict.keys():
    try:
        # Handle all cases but the first
        features = (feature_cosims_dict[feature_group]
                            .set_index(metascripts['description'])
                            .melt(var_name = 'description comparison', 
                                    value_name = f'{feature_group} similarity', 
                                    ignore_index = False)
                                .reset_index()
                                .rename(columns = {'description': 'description selection'})
                    )
        all_features = all_features.merge(features)
    except:
        # Handle the first case
        all_features = (feature_cosims_dict[feature_group]
                            .set_index(metascripts['description'])
                            .melt(var_name = 'description comparison', 
                                    value_name = f'{feature_group} similarity', 
                                    ignore_index = False)
                                .reset_index()
                                .rename(columns = {'description': 'description selection'})
                    )
# create the mean similarity column
all_features['mean similarity'] = all_features.mean(axis = 1)

### metacosims: all_features + metadata

# use all_features together with metascripts to create metacosims
metacols = ['description', 'link', 'script characters', 'id',
           'artist', 'title', 'fullTitle', 'year', 'image', 'releaseDate',
           'runtimeMins', 'runtimeStr', 'awards', 'genres',
           'companies', 'contentRating', 'imDbRating',
           'imDbRatingVotes', 'similars', 'languages']

metacosims = (all_features
                .merge(metascripts[metacols].drop(columns = 'similars'),
                       left_on = 'description selection',
                       right_on = 'description')
                .drop(columns = 'description')
                .merge(metascripts[metacols].drop(columns = 'similars'), 
                       left_on = 'description comparison', 
                       right_on = 'description',
                       suffixes = (" selection", " comparison"))
    )

# drop duplicate column
keepind = [ind for ind, col in enumerate(metacosims.columns) if ind != 1]
metacosims = metacosims.iloc[:, keepind]

# metacosims is sent to a sqlite db and pulled down into R for the Shiny dashboard

# with sqlite3.connect('../StandupRecommenderShiny/data/metacosims_recommender.sqlite') as db:
#      metacosims.to_sql('metacosims', db, if_exists = 'append', index = False)

with open('../data/metacosims_df.pickle', 'wb') as file:
    pickle.dump(metacosims, file)

### graph_all_features: all_features deduplicated for graph database upload

# create graph_all_features, a deduplicated all_features that also removes self-relations
# graph_all_features is used for Neo4J importation
graph_all_features = (all_features.loc[all_features['mean similarity'] < 0.99999]
                            .sort_values(['mean similarity', 'description selection', 'description comparison'], ascending = [False, True, True])
                            .drop_duplicates(subset = 'mean similarity', keep = 'first')
        )
graph_all_features.to_csv('../data/graph_df_all_features.csv', index = False)

### feature_mean_coms: feature groups with mean communities

# fix the indices of two feature-group dfs that were reset
# this is necessary for the upcoming concat
dfs_with_bad_indices = ['pos_props', 'pov_props']
for df in dfs_with_bad_indices:
    features_dict[df] = features_dict[df].set_index(features_dict['sur'].index)

feature_mean_coms = {group: pd.concat([df, metascripts['description']], axis = 1).merge(descom_df).rename(columns = {'community':'mean community'}) for group, df in features_dict.items()}
feature_mean_coms.keys()

feature_mean_coms['pos_props'].head()

### feature_group_coms: feature groups with group communities

import glob
communities_files = glob.glob("../data/feature_communities/*.csv")
com_file_names = [re.search("communities_(.+)\.csv", file).group(1) for file in communities_files]
com_file_names[1:]

# take all communities except bow
keys_dict = dict(zip(com_file_names[1:], sorted(list(feature_mean_coms.keys()))))

# create dictionary of feature-group community labels
feature_group_coms = {keys_dict[re.search("communities_(.+)\.csv", file).group(1)]:pd.read_csv(file) for file in communities_files[1:]}

# create feature-group dfs with group community labels
feature_coms = {key: df.merge(feature_group_coms[key]) for key, df in feature_mean_coms.items()}

feature_coms['pos_props'].head()

with open('../data/feature_communities/feature_coms_dict.pickle', 'wb') as file:
    pickle.dump(feature_coms, file)

### all_coms: all communities for each show

def merge_dfs(starting_df, dict_of_dfs):
    """Merge several dataframes together given a starting df and a dictionary of dataframes"""
    for key, df in dict_of_dfs.items():
        starting_df = starting_df.merge(df)
    merged_df = starting_df
    return merged_df

all_coms = merge_dfs(descom_df, feature_group_coms)
all_coms.head()

with open('../data/feature_communities/all_coms_df.pickle', 'wb') as file:
    pickle.dump(all_coms, file)

# appendix

# manually explore the shows in each community

def check_community_shows(com_index, show_range):
    """Return an ordered subset of shows in a selected community."""
    community_list = descom_df['community'].unique()
    com = community_list[com_index]
    sorted_subset = descom_df[descom_df['community'] == com].sort_values('description')[show_range]
    return sorted_subset

check_community_shows(com_index = 0, show_range = 40:60)

comedians_between = (metascripts.merge(descom_df)
                     .groupby('artist')
                     ['community']
                     .nunique() > 1).sum()

comedians_multiple = (metascripts.groupby('artist')['fullTitle'].count() > 1).sum()
print(f'{comedians_between} out of {comedians_multiple}, i.e, {round(100*comedians_between/comedians_multiple,2)}%, of comedians with multiple shows in the dataset had shows in different communities.')
